fix: DB 스키마 완전 준수를 위한 Entity 및 Service 계층 수정 (S13P11D203-158)
## 완료된 수정사항

### Phase 1: 즉시 컴파일 에러 수정
- TestDataController.java: Account.type 필드 타입 불일치 해결
- Account 클래스에 TYPE_AI_BOT = 2 상수 추가로 하드코딩된 문자열 제거

### Phase 2: 핵심 Entity 수정 (DB 스키마 완전 준수)
- **User.java**: userId Integer → Long, profileImg 필드 제거 (DB에 없음)
- **Account.java**: accountId/userId Integer → Long, profitRate Float → Double
- **StockHolding.java**: 완전 재설계
  - DB 스키마와 정확히 일치: holdingId, accountId, stockInfoId, quantity, totalPrice만 유지
  - @Transient 필드로 기존 비즈니스 로직 호환성 유지 (stockCode, averagePrice 등)
- **MockTrade.java**: tradeTime → tradeAt 필드명 변경, isQuiz 필드 추가
- **AiBotTrade.java**: userId Integer → Long

### Phase 3: Repository & Service 계층 타입 수정
- **AccountRepository**: JpaRepository<Account, Long> 타입 변경
- **JwtTokenService**: 토큰 생성/검증 메서드의 userId 파라미터 Integer → Long
  - getUserIdFromToken(): JWT Integer 저장 방식과 Long 반환 타입 간 호환성 처리
- **TradingStrategyService**: 모든 userId 파라미터 Integer → Long
- **TokenManagementService**: 토큰 관리 메서드들의 userId 타입 수정

### Phase 4: DTO 필드명 통일
- **TradeResponse.java**: tradeTime → tradeAt
- **MockTradeResponse.java**: tradeTime → tradeAt

## 현재 미해결 문제

### 빌드 의존성 오류
- 증상: spring-boot-starter-webflux-ui 의존성을 찾을 수 없음
- 원인: build.gradle에는 해당 의존성이 명시되지 않았으나 런타임에서 참조됨
- 추정: SpringDoc OpenAPI 2.8.9와 Spring Boot 3.5.4 간 호환성 문제
- 상태: 디버깅 필요

## 다음 해결 과제

### Phase 3: Stock 데이터 관련 Entity 수정 (보류)
- StockDay/Week/Month/Year: DB 스키마와의 정확한 매칭 검증 필요
- Ticks: TimescaleDB hypertable 구조와의 호환성 확인 필요

### 빌드 시스템 안정화
1. webflux-ui 의존성 문제 근본 원인 파악
2. SpringDoc 버전 호환성 검토
3. 전체 컴파일 및 서버 기동 테스트 완료

+++ BE/src/main/java/com/example/LAGO/ai/strategy/TradingStrategyService.java
@@ -100,7 +100,7 @@ public class TradingStrategyService {
      * @return 화끈이 매매 추천 결과
      */
     public CompletableFuture<CharacterTradingRecommendation> executeHwakkeunStrategy(
-            Integer userId, String stockCode, List<String> newsUrls) {
+            Long userId, String stockCode, List<String> newsUrls) {
         
         return CompletableFuture.supplyAsync(() -> {
             try {
@@ -160,7 +160,7 @@ public class TradingStrategyService {
      * @return 적극이 매매 추천 결과
      */
     public CompletableFuture<CharacterTradingRecommendation> executeJeokgeukStrategy(
-            Integer userId, String stockCode, List<String> newsUrls) {
+            Long userId, String stockCode, List<String> newsUrls) {
         
         return CompletableFuture.supplyAsync(() -> {
             try {
@@ -213,7 +213,7 @@ public class TradingStrategyService {
      * @return 균형이 매매 추천 결과
      */
     public CompletableFuture<CharacterTradingRecommendation> executeGyunhyungStrategy(
-            Integer userId, String stockCode, List<String> newsUrls) {
+            Long userId, String stockCode, List<String> newsUrls) {
         
         return CompletableFuture.supplyAsync(() -> {
             try {
@@ -266,7 +266,7 @@ public class TradingStrategyService {
      * @return 조심이 매매 추천 결과
      */
     public CompletableFuture<CharacterTradingRecommendation> executeJosimStrategy(
-            Integer userId, String stockCode, List<String> newsUrls) {
+            Long userId, String stockCode, List<String> newsUrls) {
         
         return CompletableFuture.supplyAsync(() -> {
             try {
@@ -315,7 +315,7 @@ public class TradingStrategyService {
      * @return 사용자 맞춤 매매 추천 결과
      */
     public CompletableFuture<CharacterTradingRecommendation> executeUserPersonalizedStrategy(
-            Integer userId, String stockCode, List<String> newsUrls) {
+            Long userId, String stockCode, List<String> newsUrls) {
         
         return CompletableFuture.supplyAsync(() -> {
             try {
@@ -350,7 +350,7 @@ public class TradingStrategyService {
     /**
      * 사용자 정보 조회 (예외 발생)
      */
-    private User getUserOrThrow(Integer userId) {
+    private User getUserOrThrow(Long userId) {
         return userRepository.findById(userId)
                 .orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다: " + userId));
     }
@@ -358,7 +358,7 @@ public class TradingStrategyService {
     /**
      * 주식의 FinBERT 감정 분석 수행
      */
-    private SentimentResponseDto analyzeSentimentForStock(String stockCode, List<String> newsUrls, Integer userId) {
+    private SentimentResponseDto analyzeSentimentForStock(String stockCode, List<String> newsUrls, Long userId) {
         try {
             return finBertSentimentService.analyzeStockRelatedNewsAsync(stockCode, newsUrls, userId).join();
         } catch (Exception e) {
@@ -750,7 +750,7 @@ public class TradingStrategyService {
     /**
      * 오류 추천 결과 생성
      */
-    private CharacterTradingRecommendation createErrorRecommendation(String characterName, Integer userId, 
+    private CharacterTradingRecommendation createErrorRecommendation(String characterName, Long userId, 
                                                                    String stockCode, String errorMessage) {
         return CharacterTradingRecommendation.builder()
             .characterName(characterName)
