+++ BE/src/main/java/com/example/LAGO/repository/MockTradeRepository.java
@@ -5,10 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface MockTradeRepository extends JpaRepository<MockTrade, Long> {
+public interface MockTradeRepository extends JpaRepository<MockTrade, Integer> {
     @Query("SELECT COUNT(mt) FROM MockTrade mt WHERE mt.accountId = :accountId")
-    Long countByAccountId(@Param("accountId") Integer accountId);
+    Long countByAccountId(@Param("accountId") Long accountId);
 
     @Query("SELECT AVG(mt.price * mt.quantity) FROM MockTrade mt WHERE mt.accountId = :accountId")
-    Double findAvgTradeValue(@Param("accountId") Integer accountId);
+    Double findAvgTradeValue(@Param("accountId") Long accountId);
 }
